requirements.txt
fastapi==0.103.0
python-dotenv==1.0.1
pydantic-settings==2.3.4
gunicorn==22.0.0
sqlalchemy==2.0.31
pymysql==1.1.1
starlette~=0.27.0
PyJWT~=2.8.0
passlib~=1.7.4
pydantic~=2.8.2
python-multipart~=0.0.9
httpx~=0.27.0
xmltodict~=0.13.0